lstm_backend/requirements.txt
absl-py==2.3.1
alpaca-py==0.42.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
astunparse==1.6.3
asyncpg==0.30.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.12.0
cycler==0.12.1
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
flatbuffers==25.2.10
fonttools==4.59.0
frozendict==2.4.6
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
keras==3.11.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.1
multitasking==0.0.12
namex==0.1.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pandas-ta==0.3.14b0
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
protobuf==5.29.5
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
scikit-learn==1.7.1
scipy==1.16.1
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sseclient-py==1.8.0
starlette==0.47.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yfinance==0.2.65
